backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
paho-mqtt>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
